refactor: global app hooks
